add all three buttons at once
